Consolidate utility functions in one file
