+++ lib/autocomplete-provider.js
@@ -1,7 +1,7 @@
 'use babel';
 
 import _ from 'lodash';
-import log from './utils/log';
+import { log } from './utils';
 import Config from './config';
 import store from './store';
