+++ lib/code-manager.js
@@ -6,7 +6,7 @@ import escapeStringRegexp from 'escape-string-regexp';
 import _ from 'lodash';
 
 import store from './store';
-import log from './utils/log';
+import { log } from './utils';
 
 export function normalizeString(code) {
   if (code) {
