+++ lib/components/inspector/index.js
@@ -3,9 +3,8 @@
 import React from 'react';
 import * as transformime from 'transformime';
 
-import log from './../../utils/log';
+import { log, reactFactory } from './../../utils';
 import store from './../../store';
-import reactFactory from './../../utils/react';
 import { getCodeToInspect } from './../../code-manager';
 import InspectorComponent from './component';
