+++ lib/kernel-manager.js
@@ -9,9 +9,8 @@ import path from 'path';
 import Config from './config';
 import ZMQKernel from './zmq-kernel';
 import KernelPicker from './kernel-picker';
-import log from './utils/log';
 import store from './store';
-import { grammarToLanguage } from './utils';
+import { grammarToLanguage, log } from './utils';
 
 
 class KernelManager {
