+++ lib/kernel.js
@@ -3,12 +3,11 @@
 import { Emitter } from 'atom';
 import { observable, action } from 'mobx';
 
-import { grammarToLanguage } from './utils';
+import { grammarToLanguage, log } from './utils';
 import store from './store';
 
 import WatchSidebar from './watch-sidebar';
 import HydrogenKernel from './plugin-api/hydrogen-kernel';
-import log from './utils/log';
 
 export default class Kernel {
   @observable executionState = 'loading'
