+++ lib/main.js
@@ -22,9 +22,7 @@ import ZMQKernel from './zmq-kernel';
 import Inspector from './components/inspector';
 import AutocompleteProvider from './autocomplete-provider';
 import HydrogenProvider from './plugin-api/hydrogen-provider';
-import reactFactory from './utils/react';
-import log from './utils/log';
-import { isMultilanguageGrammar } from './utils';
+import { log, reactFactory, isMultilanguageGrammar } from './utils';
 
 const Hydrogen = {
   config: Config.schema,
