+++ lib/result-view.js
@@ -3,7 +3,7 @@
 import { CompositeDisposable } from 'atom';
 import { createTransform } from 'transformime';
 
-import log from './utils/log';
+import { log } from './utils';
 
 const transform = createTransform();
