+++ lib/signal-list-view.js
@@ -5,7 +5,7 @@ import _ from 'lodash';
 
 import WSKernel from './ws-kernel';
 import kernelManager from './kernel-manager';
-import log from './utils/log';
+import { log } from './utils';
 import store from './store';
 
 // View to display a list of grammars to apply to the current editor.
