+++ lib/watch-view.js
@@ -3,7 +3,7 @@
 import { TextEditor } from 'atom';
 
 import ResultView from './result-view';
-import log from './utils/log';
+import { log } from './utils';
 
 export default class WatchView {
