+++ lib/watches-picker.js
@@ -2,7 +2,7 @@
 
 import { SelectListView } from 'atom-space-pen-views';
 
-import log from './utils/log';
+import { log } from './utils';
 
 class SignalListView extends SelectListView {
   initialize() {
