+++ lib/ws-kernel.js
@@ -2,7 +2,7 @@
 
 import Kernel from './kernel';
 import InputView from './input-view';
-import log from './utils/log';
+import { log } from './utils';
 
 export default class WSKernel extends Kernel {
   constructor(kernelSpec, grammar, session) {
