+++ lib/zmq-kernel.js
@@ -7,7 +7,7 @@ import v4 from 'uuid/v4';
 import Kernel from './kernel';
 import InputView from './input-view';
 import kernelManager from './kernel-manager';
-import log from './utils/log';
+import { log } from './utils';
 
 export default class ZMQKernel extends Kernel {
   constructor(kernelSpec, grammar, connection, connectionFile, kernelProcess) {
